project and tech form complete

+++ server/controllers/techs.js
@@ -1,6 +1,8 @@
 import express from 'express'
 import { Tech } from '../models/tech.js'
 import { validateTech } from '../middleware/validations.js'
+import { User } from '../models/user.js'
+import { authMiddleware } from '../middleware/authMiddleware.js'
 
 const techsRouter = express.Router()
 
@@ -18,8 +20,13 @@ techsRouter.get('/:id', async (req, res) => {
     }
 })
 
-techsRouter.post('/', validateTech, async (req, res) => {
+techsRouter.post('/', authMiddleware, validateTech, async (req, res) => {
     const { name, icon } = req.body
+    const user = await User.findById(req.user.id)
+
+    if (!user) {
+        return res.status(400).json({ error: 'user missing or not valid' })
+    }
     
     const tech = new Tech({
         name,
@@ -30,7 +37,13 @@ techsRouter.post('/', validateTech, async (req, res) => {
     res.status(201).json(savedTech)
 })
 
-techsRouter.delete('/:id', async (req, res) => {
+techsRouter.delete('/:id', authMiddleware, async (req, res) => {
+    const tech = await Tech.findById(req.params.id)
+
+    if (!tech) {
+        return res.status(404).json({ error: 'No matching technology found' })
+    }
+    
     await Tech.findByIdAndDelete(req.params.id)
     res.status(204).end()
 })
